Secrecy Module Changes

+++ database/migrations/2019_03_29_044500_create_degree_tbs_table.php
@@ -18,7 +18,7 @@ class CreateDegreeTbsTable extends Migration
             $table->integer('user_id');
             $table->string('M_Title');
             $table->string('Det1');
-            $table->mediumInteger('DegCode');
+            $table->integer('DegCode');
             $table->timestamps();
         });
     }
